Run AsyncTasks in parallel
Since API Level 11 AsyncTasks run single threaded per default. This causes some Time to wait, e. g. on the login screen, where the URL validator is still running before the login can be checked.

This change is supposed to speed up the login process.

The documentation basically says it is only careful in cases where the order of the execution is relevant, but i do not think this is a problem in our case since the AsyncTasks do not depend on each other.

https://developer.android.com/reference/android/os/AsyncTask.html